Lazy-load ZooHeader's and ZooFooter's user-specific navigation and admin feature (#5659)
* combine UserNavigation and lazy load it

* lazy load footer adminContainer
