--- packages/lib-react-components/src/ZooHeader/components/SignedInUserNavigation/index.js
@@ -1,1 +1,0 @@
-export { default } from './SignedInUserNavigation'
