--- packages/lib-react-components/src/ZooHeader/components/SignedOutUserNavigation/components/NavButton/index.js
@@ -1,1 +1,0 @@
-export { default } from './NavButton'
