--- packages/lib-react-components/src/ZooHeader/components/SignedOutUserNavigation/index.js
@@ -1,1 +1,0 @@
-export { default } from './SignedOutUserNavigation'
