+++ packages/lib-react-components/src/hooks/index.js
@@ -1,5 +1,6 @@
-export { default as useJSONData } from './useJSONData'
+export { default as useHasMounted } from './useHasMounted.js'
+export { default as useJSONData } from './useJSONData.js'
 export { default as usePanoptesUser } from './usePanoptesUser.js'
 export { default as useUnreadMessages } from './useUnreadMessages.js'
 export { default as useUnreadNotifications } from './useUnreadNotifications.js'
-export { default as useProgressiveImage } from './useProgressiveImage'
+export { default as useProgressiveImage } from './useProgressiveImage.js'
