+++ packages/lib-react-components/src/hooks/useHasMounted.js
@@ -1,0 +1,11 @@
+import { useEffect, useState } from 'react'
+
+export default function useHasMounted() {
+  const [hasMounted, setHasMounted] = useState(false)
+
+  useEffect(() => {
+    setHasMounted(true)
+  }, [])
+
+  return hasMounted
+}
